requirements.txt
sentence-transformers
uvicorn
streamlit
ctransformers
langchain
python-box